restructured alert to now be modal but has werid bug with a parsing error
